fix: 리뷰 후 반영
- Stack 안에 1depth 까지만 허용
- 변수로 분리한 배경은 다시 body 안으로 수정
- void 로직도 다시 안으로 수정
